Fix synchronization bugs Visor<->Tracer
Vulkan requires allocation to be multiple of a variable". Visor sends it to tracer and tracer gives that amount of memory.
Vulkan requires imported memory must be valid until it is released. Visor not prematurely drops the memory when camera/renderer etc. is changed. Drops frames untill new render buffer is given