[MAY BREAK] Update most dependencies
Remove BS::thread_pool as dependency (Implement in source)
Move logging functions to a *.cpp file by implementing the fmt example (immediately type erase, and do the work on compiled function)

BS::thread_pool was good but it has a problem with MSVC i think, so I've implemented it in a similar fashion.
"submit_blocks" function was somewhat broken as well (it moves the function to the wrapped std::function multiple times, which is wrong)
Our SubmitBlocks copies the callable type (i.e. Functor/Lambda) to a shared_ptr, similar to the handling of the promise. Passes this pointer
to the tasks. Since it is implemented from ground up, I've also made these shared_ptr allocations via pmr, instead of new/delete.