Add and use PackTypes variadic arg-packed struct to use instead of std::tuple
Change cub::transform_iterator to thrust::transform_iterator since it is marked as depracated in CUDA algorithms.

Conditionally check the SceneLoaderUSD.dll's availability on TracerThread since it is not a mandatory dll.

Add better exception to SharedLibrary constructor which includes the *.dll name.

[Self Note]
Report a bug on CUDA 12.9: Out-of-memory crash on the LLVM portion of the nvcc toolchain. Pinpointed the bug which is related to the alignment of Vector3f which is 16.
If it is changed to 8 crash does not occur. It does not happen all the time but it happens compiling ray casting kernels which happened to have TextureView class returning
a std::optional<Vector4> which causes the crash.