Change Tracer interface to support volumes.
Tracer now requires "Volume" (consists of "Medium" "Transform" and "Priority") to be set by surfaces or boundary surfaces.

Boundary Surfaces must give nested volume chain to correctly track media. If empty boundary surface is assumed to be in Boundary Volume.
Surfaces must give a Volume Id if they are enclosing a volume. If not these volumes assumed that are only reflective.