Properly implement/optimize medium tracker "InsertNewVolume" and "TryInsertAtomic" functions, each takes 80us for ~2M rays (given items are already in the hash table (generic case)).
Volume priorities are now signed, if it is not given for a volume, it is (INT32_MIN + 1). INT32_MIN is reserved for future use.
(i.e., we can use it to compare invalid entries later)

Add warp vote functions to the device (CUDA) which were not used. I've tried to dedicate 8-lane warp for each ray. But it had
worse performance.